requirements.txt
numpy
pillow
fastai
pathlib
pandas
sgmllib3k
torch
torchvision